Eanble warnings as errors (and fix them) and move some generated headers around

+++ Source/Mesh/RenderObject.cpp
@@ -24,7 +24,7 @@ void RenderObject::BindAndDraw(VkCommandBuffer commandBuffer, const glm::mat4& v
 
     vkCmdPushConstants(commandBuffer, material->getPipelineLayout(), VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(meshPushConstants), &meshPushConstants);
 
-    vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, material->getPipelineLayout(), 0, descriptorSets.size(), descriptorSets.data(), 0, nullptr);
+    vkCmdBindDescriptorSets(commandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, material->getPipelineLayout(), 0, static_cast<uint32_t>(descriptorSets.size()), descriptorSets.data(), 0, nullptr);
 
-    vkCmdDraw(commandBuffer, mesh->vertices.size(), 1, 0, 0);
+    vkCmdDraw(commandBuffer, static_cast<uint32_t>(mesh->vertices.size()), 1, 0, 0);
 }
